+++ android/app/src/main/java/com/zelix_rised_trades/MainActivity.java
@@ -1,0 +1,9 @@
+package com.zelix_rised_trades;
+
+import com.google.androidgamesdk.GameActivity;
+
+public class MainActivity extends GameActivity {
+    static {
+        System.loadLibrary("game_client");
+    }
+}